requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.8
django-storages==1.12
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
pyparsing==2.4.7
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
gunicorn==20.0.4
uritemplate==4.0.0
urllib3==1.26.7
six==1.15.0
Pillow==8.0.1
beautifulsoup4==4.11.1
django-cors-headers==3.13.0
requests-cache==0.9.7
sendgrid==6.9.2
boto3==1.16.22
psycopg2-binary==2.9.5
seaborn==0.13.2
numpy==2.0.2
openpyxl==3.1.5
plotly==6.0.0
matplotlib==3.9.4
pandas==2.2.3